qcomputing-v0/requirements.txt
qiskit==0.20.0
qiskit-aer==0.6.1
qiskit-aqua==0.7.5
qiskit-ibmq-provider==0.8.0
qiskit-ignis==0.4.0
qiskit-terra==0.15.1
pandas==1.0.3
numpy==1.17.2
matplotlib==3.2.1
Pillow==2.2.1
imageloader==0.0.5
